app/requirements.txt
alembic==1.5.4
bcrypt==3.2.0
cffi==1.14.4
click==7.1.2
cryptography==3.3.1
ecdsa==0.14.1
fastapi==0.63.0
h11==0.12.0
Mako==1.1.4
MarkupSafe==1.1.1
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.20
pydantic==1.7.3
python-dateutil==2.8.1
python-editor==1.0.4
python-jose==3.2.0
python-multipart==0.0.5
rsa==4.7
six==1.15.0
SQLAlchemy==1.3.23
SQLAlchemy-Utils==0.36.8
starlette==0.13.6
uvicorn==0.13.3
